Fixed checksum computation. Now computes correct one's complement.

+++ check.h
@@ -1,3 +1,3 @@
 #include <stdint.h>
 
-int csum(uint8_t *addr, int count);
+int csum(uint16_t *addr, int count);
